Refactored modules
Cleaner design, easier to test and build on
