+++ lib/Modules/modules.h
@@ -1,0 +1,8 @@
+#include "cut_the_wire.h"
+#include "keypad.h"
+#include "simon_says.h"
+#include "speaker.h"
+#include "strike.h"
+#include "timer.h"
+
+#pragma once
