--- lib/scheduler.cpp
@@ -1,8 +1,0 @@
-//Handler for scheduled tasks e.g: beep
-class ScheduledHandler{
-    
-    public:
-    void init_scheduled_handler(){
-
-    }
-};
